Moving increment_filename function into its own file since so many files used it
